requirements.txt
altgraph==0.17.4
click==8.1.7
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
mypy==1.11.1
mypy-extensions==1.0.0
packaging==24.1
pefile==2023.2.7
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pywin32-ctypes==0.2.3
rich==13.7.1
setuptools==73.0.1
typer-slim==0.12.4
types-pyinstaller==6.10.0.20240812
typing_extensions==4.12.2
